Moved V0 and V1 User/Role/Permision endpoints from the users app to the api app
