+++ api/app/signals/apps/api/v0/views.py
@@ -8,6 +8,7 @@ from datapunt_api.pagination import HALPagination
 from datapunt_api.rest import DatapuntViewSet
 from django.conf import settings
 from django.core.exceptions import ValidationError as DjangoCoreValidationError
+from django.http import JsonResponse
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import mixins, viewsets
 from rest_framework.exceptions import ValidationError
@@ -240,3 +241,34 @@ class MlPredictCategoryView(APIView):
                     data[key].append([response_data[key][1][0]])
 
         return Response(data)
+
+
+class UserMeView(APIView):
+    """Handle information about user me."""
+    authentication_classes = (JWTAuthBackend, )
+
+    def get(self, request):
+        data = {
+            'username': request.user.username,
+            'email': request.user.email,
+            'is_staff': request.user.is_staff is True,
+            'is_superuser': request.user.is_superuser is True,
+            'permissions': [
+                permission
+                for permission in request.user.get_all_permissions()
+                if permission.startswith('signals')
+            ],
+        }
+
+        groups = []
+        departments = []
+        for group in request.user.groups.all():
+            match = re.match(r"^dep_(\w+)$", group.name)
+            if match:
+                departments.append(match.group(1))
+            else:
+                groups.append(group.name)
+        data['groups'] = groups
+        data['departments'] = departments
+
+        return JsonResponse(data)
