+++ api/app/signals/apps/api/v1/views/__init__.py
@@ -16,12 +16,15 @@ from signals.apps.api.v1.views.departments import PrivateDepartmentViewSet
 # from signals.apps.api.v1.views.ml_tool_proxy import MlPredictCategoryView  # V1 disabled for now
 from signals.apps.api.v1.views.namespace import NamespaceView
 from signals.apps.api.v1.views.pdf import GeneratePdfView
+from signals.apps.api.v1.views.permission import PermissionViewSet
 from signals.apps.api.v1.views.questions import PublicQuestionViewSet
+from signals.apps.api.v1.views.role import RoleViewSet
 from signals.apps.api.v1.views.signal import PrivateSignalViewSet, PublicSignalViewSet
 from signals.apps.api.v1.views.signal_split import PrivateSignalSplitViewSet
 from signals.apps.api.v1.views.source import PrivateSourcesViewSet
 from signals.apps.api.v1.views.status_message_template import StatusMessageTemplatesViewSet
 from signals.apps.api.v1.views.stored_signal_filter import StoredSignalFilterViewSet
+from signals.apps.api.v1.views.user import UserViewSet
 
 __all__ = (
     'PublicSignalAttachmentsViewSet',
@@ -43,4 +46,7 @@ __all__ = (
     'PublicAreasViewSet',
     'PrivateAreasViewSet',
     'PrivateSourcesViewSet',
+    'PermissionViewSet',
+    'RoleViewSet',
+    'UserViewSet',
 )
