+++ api/app/signals/apps/api/v1/views/role.py
@@ -4,7 +4,7 @@ from django.db.models.functions import Lower
 from rest_framework.permissions import DjangoModelPermissions
 
 from signals.apps.api.generics.permissions import SIAPermissions
-from signals.apps.users.v1.serializers import RoleSerializer
+from signals.apps.api.v1.serializers import RoleSerializer
 from signals.auth.backend import JWTAuthBackend
