--- api/app/signals/apps/users/v1/serializers/__init__.py
@@ -1,13 +1,0 @@
-from signals.apps.users.v1.serializers.permission import PermissionSerializer
-from signals.apps.users.v1.serializers.profile import ProfileDetailSerializer, ProfileListSerializer
-from signals.apps.users.v1.serializers.role import RoleSerializer
-from signals.apps.users.v1.serializers.user import UserDetailHALSerializer, UserListHALSerializer
-
-__all__ = [
-    'PermissionSerializer',
-    'ProfileListSerializer',
-    'ProfileDetailSerializer',
-    'RoleSerializer',
-    'UserDetailHALSerializer',
-    'UserListHALSerializer',
-]
