--- api/app/signals/apps/users/v1/views/__init__.py
@@ -1,9 +1,0 @@
-from signals.apps.users.v1.views.permission import PermissionViewSet
-from signals.apps.users.v1.views.role import RoleViewSet
-from signals.apps.users.v1.views.user import UserViewSet
-
-__all__ = [
-    'PermissionViewSet',
-    'RoleViewSet',
-    'UserViewSet',
-]
